+++ src/cassandra/statement.rs
@@ -445,7 +445,7 @@ impl Statement {
         unsafe {
             let timeout_millis = match timeout {
                 None => CASS_UINT64_MAX as u64,
-                Some(time) => time.num_milliseconds() as u64,
+                Some(time) => time.whole_milliseconds() as u64,
             };
             cass_statement_set_request_timeout(self.inner(), timeout_millis);
         }