+++ CHANGELOG.md
@@ -23,6 +23,7 @@ version number is tracked in the file `VERSION`.
   requirement that the `CassResult` must live longer than the `Row`.
 - Switched to using `parking_lot::Mutex` instead of `std::sync::Mutex` for
   `CassFuture` coordination.
+- Implemented `size_hint` on `ResultIterator`.
 
 ### Fixed
  - `CassResult::set_paging_state_token` was implemented incorrectly, namely, it did nothing,